pyproject.toml
[project]
name = "mars-daemons"
version = "0.4.0"
description = "Standalone agents backend built from the exportable agents_v2 core."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Pedro Rios", email = "priosabarca@gmail.com" },
]
keywords = [
    "agent",
    "llm",
    "claude",
    "anthropic",
    "openai",
    "runtime",
    "daemon",
    "enterprise",
    "coding-agent",
    "mcp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "anthropic>=0.40",
    "openai>=1.0",
    "litellm>=1.74",
    "httpx>=0.28",
    "pydantic>=2.6",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "sse-starlette>=2.1",
    "anyio>=4.8",
    "python-multipart>=0.0.9",
]

[project.urls]
Homepage = "https://github.com/TacosyHorchata/mars-daemons"
Repository = "https://github.com/TacosyHorchata/mars-daemons"
Issues = "https://github.com/TacosyHorchata/mars-daemons/issues"
Changelog = "https://github.com/TacosyHorchata/mars-daemons/blob/main/CHANGELOG.md"

[project.scripts]
mars-daemons = "mars_runtime.cli.run:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mars_runtime"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
addopts = "-q"
